Rename image archive directories and document the lossy limitation
The Python package did not contain images and the images/ directory
contained no images. Rename to reflect what each actually is:

  oasis_images/ -> image_archive/        the dataset-agnostic tool
  images/*      -> image_archive/axiom/  one worked instance of it

Test modules follow as tests/test_image_archive_*.py, and the package
docstrings drop the OASIS-specific wording where the code is generic,
keeping "Axiom-style" only where the six-column index schema is a real
constraint.

The run receipt is left intact. Its command arrays, contract_path, and
runtime package name record what actually executed on 4-5 Aug, so
rewriting them would falsify the record. A [layout_rename] section maps
old paths to new instead. source.toml moved unmodified, so the receipt's
contract_sha256 still binds.

Also add a Known limitation section to the archive README. The tier is
lossy and no one has checked whether profiles computed from the .jxl
derivative match profiles from the source TIFFs. Measured for context:
the archive is 22-88x smaller than lossless JPEG XL of the same pixels,
and Brightfield error is 4-8x worse than any fluorescence channel
(3.02% RMSE of signal range vs 0.38-0.75%), consistently across all four
batches, because transmitted light occupies a narrow band that a
perceptual metric does not know is load-bearing.

Verified: 31 tests plus 7 subtests, ruff, and pyright pass; the renamed
CLI reports ledger_complete against the live archive and its preflight
gate resolves through the new contract path.
