+++ image_archive/axiom/oasis-axiom-jpegxl.service
@@ -12,7 +12,7 @@ UMask=0007
 Nice=10
 IOSchedulingClass=best-effort
 IOSchedulingPriority=6
-ExecStart=/run/current-system/sw/bin/nix develop path:. --command pixi run -e images python -m oasis_images archive --contract images/source.toml --workers 64 --max-in-flight 128 --max-attempts 5 --max-consecutive-failures 32
+ExecStart=/run/current-system/sw/bin/nix develop path:. --command pixi run -e images python -m image_archive archive --contract image_archive/axiom/source.toml --workers 64 --max-in-flight 128 --max-attempts 5 --max-consecutive-failures 32
 KillSignal=SIGINT
 TimeoutStopSec=10min
 Restart=on-abnormal
